Uses column number_of_recipes
Instead of being able to use the column number_of_recipes to represent how many
of each recipe are in a meal plan, there is a meal plan recipe record per
recipe, regardless if it is a duplicate record. Amends queries to multiply by
number_of_recipes so that duplicat records are no longer necessary.