Adds include_blank for ingredient form
When an ingredient is added or updated without selecting a unit for its
measuring amount, the record is created with the first unit option as its unit
without indicating on the form. This happens because the units are selected
using a dropdown that does not have the html option, 'include_blank'. Adds this
option since an ingredient record doesn't need to be tied to a unit.