Deletes creating recipe names from calender mealplan (#152)
Prefer to manually create recipe name if it is indeed necessary to do so.
Sometimes it is just an error in naming, and the true recipe does exist.

Refactors logic in creating MealPlanRecipes from calendar mealplan
Consolidates the amount of times having to loop over meal plan info and having
lots of methods to create the necessary records at the cost of hitting db in a
loop. Benefit is that code is more readable and in the end, the net performance
has not been severely negatively affected.